+++ pwsafe/pwsafe/corelib/PWCharPool.h
@@ -55,8 +55,8 @@ class CPasswordCharPool
  private:
   enum CharType {LOWERCASE = 0, UPPERCASE = 1,
                  DIGIT = 2, SYMBOL = 3, HEXDIGIT = 4, NUMTYPES = 5};
-  CharType GetRandomCharType(unsigned int rand) const; // select a chartype with weighted probability
-  TCHAR GetRandomChar(CharType t, unsigned int rand) const;
+  CharType GetRandomCharType(size_t rand) const; // select a chartype with weighted probability
+  TCHAR GetRandomChar(CharType t, size_t rand) const;
 
   // here are all the character types, in both full and "easyvision" versions
   static const TCHAR std_lowercase_chars[];
