+++ pwsafe/pwsafe/corelib/PWSclipboard.cpp
@@ -9,7 +9,6 @@
 /** \file
  * Implementation file for the PWSclipboard class
  */
-#include "../stdafx.h"
 
 #include <afxole.h>
 #include "PWSclipboard.h"
@@ -63,7 +62,7 @@ PWSclipboard::ClearData()
     HANDLE hData = odo.GetGlobalData(CLIPBOARD_TEXT_FORMAT);
     if (hData != NULL) {
       LPCTSTR pData = (LPCTSTR)::GlobalLock(hData);
-      SIZE_T dwlength =  ::GlobalSize(hData) - sizeof(TCHAR); // less trailing null
+      DWORD dwlength =  ::GlobalSize(hData) - sizeof(TCHAR); // less trailing null
       if (dwlength < 1)
         return !m_set;
